refactor(label_table): extract LabelRemap class and remove Hole functionality
- Create independent LabelRemap module with full Doxygen documentation
- Remove Hole functionality (hole_list_, hole_mutex_, 5 methods)
- Simplify hgraph ID allocation to pure increment strategy
- Remove Hole-related tests from label_table_test.cpp
- Add src/impl/label_table/CMakeLists.txt for new directory structure

Fixes: #2088
