+++ src/algorithm/brute_force.h
@@ -17,7 +17,7 @@
 
 #include "algorithm/inner_index_interface.h"
 #include "brute_force_parameter.h"
-#include "impl/label_table.h"
+#include "impl/label_table/label_table.h"
 #include "typing.h"
 #include "utils/pointer_define.h"
 #include "vsag/filter.h"
