+++ src/impl/CMakeLists.txt
@@ -27,16 +27,17 @@ add_subdirectory (odescent)
 add_subdirectory (reorder)
 add_subdirectory (blas)
 add_subdirectory (reasoning)
+add_subdirectory (label_table)
 
 file (GLOB IMPL_SRCS "*.cpp")
 list (FILTER IMPL_SRCS EXCLUDE REGEX "_test.cpp")
 
 add_library (impl OBJECT ${IMPL_SRCS})
 target_link_libraries (impl PRIVATE transform thread_pool allocator heap vsag_blas
-    bitset logger filter cluster searcher odescent reorder reasoning coverage_config vsag_src_common)
+    bitset logger filter cluster searcher odescent reorder reasoning label_table coverage_config vsag_src_common)
 
 set (IMPL_LIBS transform thread_pool allocator heap bitset vsag_blas
-    logger filter impl cluster searcher odescent reorder reasoning PARENT_SCOPE)
+    logger filter impl cluster searcher odescent reorder reasoning label_table PARENT_SCOPE)
 
 if (ENABLE_TESTS)
     file (GLOB_RECURSE IMPL_TESTS "*_test.cpp")
