+++ src/impl/label_table/CMakeLists.txt
@@ -1,0 +1,25 @@
+
+# Copyright 2024-present the vsag project
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+set (LABEL_TABLE_SRC
+        label_remap.cpp
+        label_remap.h
+        label_table.cpp
+        label_table.h
+)
+
+add_library (label_table OBJECT ${LABEL_TABLE_SRC})
+target_link_libraries (label_table PRIVATE coverage_config vsag_src_common)
